Null public_key should be accepted if undefined
Normally, patchy fields will not delete the field if you do not pass it
through PUT. However, if you pass it a "null" in the JSON, then the API
will set the field to null.

However, the various clients will sometimes send "null" for public_key,
and expect the API to preserve the public_key in the database. These
changes will test this idiosyncratic behavior of public_key.