Add minimal tests for depsolving with constraints in an environment
Needed to address CHEF-3699. The environment used for most of the
tests now includes some cookbook version constraints. In addition, a
strict environment that excludes the testing cookbooks is added to
validate that the environment constraints are respected.

Considerably more coverage for happy path and error condition cases
could be added to this spec. The environments/cookbooks_spec.rb file
contains setup code that could be repurposed.

The error messages in this patch may need to be adjusted for use
against a Ruby-based Chef server.