Change the default pedant config location
The old default was where we used to keep the config on Private Chef.  This obviously isn't too useful for Open Source, and in any case, we've again changed where the config is stored in our Omnibus builds.

This will now default to a "pedant_config.rb" file in the current directory, as a sensible compromise.

(An alternative, of course, would be to make the config file a mandatory argument, and not provide a default at all.)