Re-implemented Hash matcher
This should be more robust, allowing both strict AND loose matching to
fully cascade down a series of nested hashes.

The logic is a little cleaner now, since we're comparing two hashes
directly, instead of on a key/value pair basis.

As a bonus, the failure messages should now be easier to interpret, as
both the actual hash and the expected "spec" hash are pretty-printed.

Tests, too!