Expect missing header value to behave same as missing header
By RFC2616, HTTP header values do not include leading or trailing
white space. Therefore, a server should treat a header with no value
the same as when that header is not present.

This means that we should expect 400 response codes when sending empty
headers just as we do when omitting an authorization header (rather
than a 401).

This change was motivated by the latest version of mochiweb (used by
latest version of webmachine) which includes a change such that the
header access function returns 'undefined' when either the header is
not present or contains only white space.

The tests for this have now been marked pending for ruby endpoints to
avoid backporting this behavior.