CHEF-4086: Explicitly use default_port in the server url
Since erchef now uses the value of Hosts to create urls returned back to
the client, this value will have a port, regardless of whether it is a
default port or not.

All Pedant tests that tests these returned values constructs the
expected URL off of chef_server_url (in the Pedant config file). This
patch explicitly adds the port to the url even if it is a default port.