this needs to be relaxed to 1.6.7
actual chef-client has a restriction on rest-client <= 1.6.7

the ipv6 hacked up opscode/rest-client version has a branch
which lies about its version number and sets it to 1.6.7 so that
it can be loaded along with chef-client.

getting the ipv6 patches for rest-client accepted upstream would not
help now since 1.7.0 brings in rdoc which breaks our builds on solaris.