Fixed deprecation warning for environments tests
There was an attempt to keep this in before(:all). However, since the
macros are shared with let() based tests, this created a number of
failing tests. This commit converts many of thoses tests to
before(:each) block, though it slows down the tests since it has to
setup cookbooks for each individual test. We can circle back around
later to attempt to make this a before(:all) block, though it might not
be possible.